requirements.txt
streamlit
langchain
langchain-openai
langchain-google-genai
google-generativeai
python-dotenv
nest_asyncio
tiktoken